requirements.txt
Cython
pyjnius